Implemented new sound engine (WIP)

+++ rust/pyxel-wrapper/src/channel_wrapper.rs
@@ -46,7 +46,7 @@ impl Channel {
     pub fn play(
         &self,
         snd: Bound<'_, PyAny>,
-        tick: Option<u32>,
+        tick: Option<f64>,
         r#loop: Option<bool>,
         resume: Option<bool>,
     ) -> PyResult<()> {
